add data_loading.py
